refactor(quantize): extract graph rewrite and weight math
Move bump_opset / MatMul-Conv integer chains into graph.rs and
per-channel quantization into weights.rs. Public quantize_model
is unchanged.
